@@ -4,12 +4,10 @@ import App from './components/app/App';
 
 import './style/style.scss';
 
-
-
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-
+ReactDOM
+    .createRoot(document.getElementById('root'))
+    .render(
+        <React.StrictMode>
+            <App />
+        </React.StrictMode>,
+    );
