@@ -1,43 +1,28 @@
-import {useState} from "react";
-
+import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-
-import decoration from '../../resources/img/vision.png';
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-import AppBanner from "../appBanner/AppBanner";
-import ComicsList from "../comicsList/ComicsList";
+import MainPage from "../pages/MainPage";
+import ComicsPage from "../pages/ComicsPage";
 
 const App = () => {
 
-    const [selectedChar, setChar] = useState(null);
 
-    const onCharSelected = (id) => {
-        setChar(id);
-    }
 
     return (
-        <div className="app">
-            <AppHeader/>
-            <main>
-                <AppBanner/>
-                <ComicsList/>
-                {/*<ErrorBoundary>*/}
-                {/*    <RandomChar/>*/}
-                {/*</ErrorBoundary>*/}
-                {/*<div className="char__content">*/}
-                {/*    <ErrorBoundary>*/}
-                {/*        <CharList onCharSelected={onCharSelected}/>*/}
-                {/*    </ErrorBoundary>*/}
-                {/*    <ErrorBoundary>*/}
-                {/*        <CharInfo charId={selectedChar}/>*/}
-                {/*    </ErrorBoundary>*/}
-                {/*</div>*/}
-                <img className="bg-decoration" src={decoration} alt="vision"/>
-            </main>
-        </div>
+        <Router>
+            <div className="app">
+                <AppHeader/>
+                <main>
+                    <Switch>
+                        <Route exact path='/comics'>
+                            <ComicsPage/>
+                        </Route>
+                        <Route exact path='/'>
+                            <MainPage/>
+                        </Route>
+                    </Switch>
+                </main>
+            </div>
+        </Router>
     )
 
 }
